fix(striker-ui): apply duplicate test to prefix and sequence in manifest forms
